refactor(components/atom/backToTop): add tests
