+++ components/atom/backToTop/src/config.js
@@ -1,0 +1,16 @@
+export const BASE_CLASS = 'sui-AtomBackToTop'
+export const CLASS_ICON = `${BASE_CLASS}-icon`
+export const CLASS_TEXT = `${BASE_CLASS}-text`
+export const CLASS_SHOW = `${BASE_CLASS}--show`
+export const CLASS_HIDE = `${BASE_CLASS}--hide`
+export const CLASS_READY = `${BASE_CLASS}--ready`
+
+export const SCROLL_BEHAVIOR = {
+  SMOOTH: 'smooth',
+  AUTO: 'auto'
+}
+
+export const STYLES = {
+  DARK: 'dark',
+  LIGHT: 'light'
+}
