--- components/atom/backToTop/src/settings.js
@@ -1,9 +1,0 @@
-export const SCROLL_BEHAVIOR = {
-  SMOOTH: 'smooth',
-  AUTO: 'auto'
-}
-
-export const STYLES = {
-  DARK: 'dark',
-  LIGHT: 'light'
-}
