+++ components/atom/backToTop/test/index.test.js
@@ -11,51 +11,148 @@ import chai, {expect} from 'chai'
 import chaiDOM from 'chai-dom'
 import {createRef} from 'react'
 
+import * as pkg from '../src'
+
+import json from '../package.json'
+
 chai.use(chaiDOM)
 
-describe('atom/backToTop', () => {
-  const Component = AtomBackToTop
+describe(json.name, () => {
+  const {default: Component} = pkg
   const setup = setupEnvironment(Component)
 
-  it('should render without crashing', () => {
+  it('library should include defined exported elements', () => {
     // Given
-    const props = {}
+    const library = pkg
+    const libraryExportedMembers = [
+      'backToTopStyles',
+      'backToTopScrollBehavior',
+      'default'
+    ]
 
     // When
-    const component = <Component {...props} />
+    const {
+      backToTopStyles,
+      backToTopScrollBehavior,
+      default: AtomBackToTop,
+      ...others
+    } = library
 
     // Then
-    const div = document.createElement('div')
-    ReactDOM.render(component, div)
-    ReactDOM.unmountComponentAtNode(div)
+    expect(Object.keys(library).length).to.equal(libraryExportedMembers.length)
+    expect(Object.keys(library)).to.have.members(libraryExportedMembers)
+    expect(Object.keys(others).length).to.equal(0)
   })
 
-  it('should NOT render null', () => {
-    // Given
-    const props = {}
-
-    // When
-    const {container} = setup(props)
-
-    // Then
-    expect(container.innerHTML).to.be.a('string')
-    expect(container.innerHTML).to.have.lengthOf(0)
-  })
-
-  describe('forwardRef', () => {
-    it('should return forwardRef html button element when giving a ref to the component', () => {
+  describe(Component.displayName, () => {
+    it('should render without crashing', () => {
       // Given
-      const props = {children: 'button'}
-      const ref = createRef()
+      const props = {}
 
       // When
-      const component = <Component {...props} ref={ref} />
-      const div = document.createElement('div')
-      ReactDOM.render(component, div)
+      const component = <Component {...props} />
 
       // Then
-      expect(ref.current).to.not.equal(undefined)
-      expect(ref.current.nodeName).to.equal('BUTTON')
+      const div = document.createElement('div')
+      ReactDOM.render(component, div)
+      ReactDOM.unmountComponentAtNode(div)
+    })
+
+    it('should NOT render null', () => {
+      // Given
+      const props = {}
+
+      // When
+      const {container} = setup(props)
+
+      // Then
+      expect(container.innerHTML).to.be.a('string')
+      expect(container.innerHTML).to.have.lengthOf(0)
+    })
+
+    describe('forwardRef', () => {
+      it('should return forwardRef html button element when giving a ref to the component', () => {
+        // Given
+        const props = {children: 'button'}
+        const ref = createRef()
+
+        // When
+        const component = <Component {...props} ref={ref} />
+        const div = document.createElement('div')
+        ReactDOM.render(component, div)
+
+        // Then
+        expect(ref.current).to.not.equal(undefined)
+        expect(ref.current.nodeName).to.equal('BUTTON')
+      })
+    })
+  })
+
+  describe('backToTopStyles', () => {
+    it('value must be an object enum', () => {
+      // Given
+      const library = pkg
+
+      // When
+      const {backToTopStyles: actual} = library
+
+      // Then
+      expect(actual).to.be.an('object')
+    })
+
+    it('value must be a defined string-key pair filled', () => {
+      // Given
+      const library = pkg
+      const expected = {
+        DARK: 'dark',
+        LIGHT: 'light'
+      }
+
+      // When
+      const {backToTopStyles: actual} = library
+      const {DARK, LIGHT, ...others} = actual
+
+      // Then
+      expect(Object.keys(others).length).to.equal(0)
+      expect(Object.keys(actual)).to.have.members(Object.keys(expected))
+      Object.entries(expected).forEach(([expectedKey, expectedValue]) => {
+        expect(Object.keys(actual).includes(expectedKey)).to.be.true
+        expect(actual[expectedKey]).to.equal(expectedValue)
+      })
+    })
+  })
+
+  describe('backToTopScrollBehavior', () => {
+    it('value must be an object enum', () => {
+      // Given
+      const library = pkg
+
+      // When
+      const {backToTopScrollBehavior: actual} = library
+
+      // Then
+      expect(actual).to.be.an('object')
+    })
+
+    it('value must be a defined string-key pair filled', () => {
+      // Given
+      const library = pkg
+      const expected = {
+        SMOOTH: 'smooth',
+        AUTO: 'auto'
+      }
+
+      // When
+      const {backToTopScrollBehavior: actual} = library
+      const {SMOOTH, AUTO, ...others} = actual
+
+      // Then
+      expect(Object.keys(others).length).to.equal(0)
+      expect(Object.keys(actual)).to.have.members(Object.keys(expected))
+      Object.entries(expected).forEach(([expectedKey, expectedValue]) => {
+        expect(Object.keys(actual).includes(expectedKey)).to.be.true
+        expect(actual[expectedKey]).to.equal(expectedValue)
+      })
     })
   })
 })
